Refactor project structure
- Changed the main entry point from `main.js` to `index.js`.
- Removed obsolete core modules: `assets.js`, `gameState.js`, and `storage.js`.
- Introduced new manager modules: `assetsManager.js`, `inputManager.js`, `mobileControlsManager.js`, `playerManager.js`, `settingsMenuManager.js`, `setupManager.js`, `stateManager.js`, and `storageManager.js` to handle respective functionalities.
- Created a new `game.js` file to initialize the game and manage the game loop.
- Updated player controls and interactions to utilize the new manager structure.
- Implemented a new HUD module to display game information.
- Added mobile control setup and interaction logic.
- Refactored map generation and tile creation logic into dedicated modules.
- Enhanced the instruction overlay for better user experience.
- Cleaned up unused files and organized the codebase for maintainability.
