+++ index.html
@@ -16,7 +16,7 @@
 </head>
 
 <body>
-  <script type="module" src="/src/main.js"></script>
+  <script type="module" src="/src/index.js"></script>
   <noscript>You need to enable JavaScript to run this app.</noscript>
 </body>
