+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "pixel-realm",
   "private": true,
-  "version": "0.12.6",
+  "version": "0.12.7",
   "type": "module",
   "scripts": {
     "dev": "vite",
