+++ src/manager/playerManager.js
@@ -1,0 +1,3 @@
+export { createPlayer } from "../player/factory/playerFactory.js";
+export { updatePlayerMovement } from "../player/controls/playerMovement.js";
+export { TILE_TYPES, TILE_NAMES } from "../world/tiles/tiles.js";
