+++ src/manager/worldManager.js
@@ -1,0 +1,2 @@
+export  {generateMap} from "../world/generation/generateMap.js";
+
