+++ src/player/camera/playerCamera.js
@@ -1,4 +1,4 @@
-import { CAM_DISTANCE, CAM_HEIGHT } from "./playerConstants.js";
+import { CAM_DISTANCE, CAM_HEIGHT } from "../player/playerConstants.js";
 
 export function updateCamera(camera, player, rotation) {
   camera.position.x = player.position.x + Math.sin(rotation) * CAM_DISTANCE;
