--- src/player/player.js
@@ -1,3 +1,0 @@
-export { createPlayer } from "./player/playerFactory.js";
-export { updatePlayerMovement } from "./player/playerMovement.js";
-export { TILE_TYPES, TILE_NAMES } from "./player/playerConstants.js";
