+++ src/player/player/playerConstants.js
@@ -1,17 +1,3 @@
-export const TILE_TYPES = {
-  grass: 0,
-  dirt: 1,
-  sand: 2,
-  water: 3,
-};
-
-export const TILE_NAMES = {
-  0: "grass",
-  1: "dirt",
-  2: "sand",
-  3: "water",
-};
-
 export const STRAFE_THRESHOLD = 15;
 export const SPEED = 0.05;
 export const ROTATION_SPEED = 0.03;
