+++ src/ui/settingsMenu/gameOptions.js
@@ -1,10 +1,10 @@
-import { storage } from "../../core/storage.js";
+import { storageManager } from "../../manager/storageManager.js";
 import { createMenuButton } from "./menuButton.js";
 
 export function createResetButton() {
   return createMenuButton("Reset World", () => {
     if (confirm("Reset the world? This will clear all your changes.")) {
-      storage.clearSavedData();
+      storageManager.clearSavedData();
       location.reload();
     }
   });
