+++ src/world/tiles/tiles.js
@@ -1,0 +1,13 @@
+export const TILE_TYPES = {
+  grass: 0,
+  dirt: 1,
+  sand: 2,
+  water: 3,
+};
+
+export const TILE_NAMES = {
+  0: "grass",
+  1: "dirt",
+  2: "sand",
+  3: "water",
+};
