--- src/world/world.js
@@ -1,2 +1,0 @@
-export  {generateMap} from "./generateMap.js";
-
